2to3: Update module syntax for running Python3 (#1)
Only a few changes were required in this module to be Python3-compliant.
Unfortuantely it remains to deal with `prctl` as that library fails on
the import since it uses a print statement. The library is unmaintained.

We can switch to `python-pctl` but that will require help to install on
our sandboxes and on development. It needs extra development packages to
be installed before it can be `pip install`'ed or `easy_install`'ed.